db changed && implemented - firebase firestore removed from project - gorm orm added - postgres db configured for usage - handlers edited to use gorm - docker compose edited to include postgres db version 17

+++ handlers/info/type.go
@@ -12,10 +12,10 @@ func NewInfoHandler() interfaces.InfoInterface {
 
 type tvStats struct {
 	amount    int
-	totalTime int
+	totalTime int64
 }
 
 type movieStats struct {
 	amount    int
-	totalTime int
+	totalTime int64
 }
